Fix Thuc Don

+++ text/DAO/FoodDAO.cs
@@ -72,7 +72,7 @@ namespace text.DAO
             return list;
         }
         
-        public bool InsertFood(string name, int id, float price)
+        public bool InsertFood(string name, int id, decimal price)
         {
             string query = string.Format("INSERT dbo.Food ( name, idCategory, price )VALUES  ( N'{0}', {1}, {2})", name, id, price);
             int result = DataProvider.Instance.ExecuteNonQuery(query);
@@ -80,7 +80,7 @@ namespace text.DAO
             return result > 0;
         }
 
-        public bool UpdateFood(int idFood, string name, int id, float price)
+        public bool UpdateFood(int idFood, string name, int id, decimal price)
         {
             string query = string.Format("UPDATE dbo.Food SET name = N'{0}', idCategory = {1}, price = {2} WHERE id = {3}", name, id, price, idFood);
             int result = DataProvider.Instance.ExecuteNonQuery(query);
